backend/src/services/livecc-inference/app/requirements.txt
livekit-server-sdk-python==1.0.0
livekit-agents==0.5.0
python-dotenv==1.0.0
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
numpy==1.26.3
pillow==10.2.0
requests==2.31.0
